refactor: move feature details to the respective page
