+++ src/needs-refactoring/components/CombinedFeaturePanel/CombinedFeaturePanel.tsx
@@ -2,14 +2,14 @@ import { Box, Callout } from "@radix-ui/themes";
 import { t } from "@transifex/native";
 import { uniqBy } from "lodash";
 import React from "react";
-import { useExpertMode } from "~/needs-refactoring/lib/useExpertMode";
+import ErrorBoundary from "~/needs-refactoring/components/shared/ErrorBoundary";
 import {
   type AnyFeature,
   isSearchResultFeature,
 } from "~/needs-refactoring/lib/model/geo/AnyFeature";
-import ErrorBoundary from "~/needs-refactoring/components/shared/ErrorBoundary";
+import { useExpertMode } from "~/needs-refactoring/lib/useExpertMode";
+import FeatureDetails from "~/pages/[placeType]/[id]/_components/FeatureDetails";
 import FeaturesDebugJSON from "./components/FeaturesDebugJSON";
-import FeatureDetails from "~/modules/feature-panel/components/FeatureDetails";
 
 type Props = {
   features: AnyFeature[];
