+++ src/needs-refactoring/components/CombinedFeaturePanel/components/AccessibilitySection/EditButton.tsx
@@ -1,9 +1,9 @@
-import { Pencil } from "lucide-react";
 import { IconButton } from "@radix-ui/themes";
 import { t } from "@transifex/native";
+import { Pencil } from "lucide-react";
 
 import React, { useContext, useState } from "react";
-import { AutoEditor } from "~/modules/edit/components/AutoEditor";
+import { AutoEditor } from "~/pages/[placeType]/[id]/_components/AutoEditor";
 import { FeaturePanelContext } from "../../FeaturePanelContext";
 
 export function EditButton({
