+++ src/pages/[placeType]/[id]/_components/AutoEditor.tsx
