+++ src/pages/[placeType]/[id]/_components/BaseEditor.tsx
