+++ src/pages/[placeType]/[id]/_components/FeatureDescription.tsx
