+++ src/pages/[placeType]/[id]/_components/HeaderImageSection.tsx
