+++ src/pages/[placeType]/[id]/_components/PartOf.tsx
