+++ src/pages/[placeType]/[id]/_components/StringFieldEditor.tsx
