+++ src/pages/[placeType]/[id]/_components/ToiletRadioCards.tsx
