+++ src/pages/[placeType]/[id]/_components/ToiletsWheelchairEditor.tsx
@@ -2,14 +2,14 @@ import { Dialog, Flex } from "@radix-ui/themes";
 import { t } from "@transifex/native";
 import type React from "react";
 import { useEffect, useState } from "react";
-import type { YesNoUnknown } from "~/needs-refactoring/lib/model/ac/Feature";
-import { isOrHasAccessibleToilet } from "~/needs-refactoring/lib/model/accessibility/isOrHasAccessibleToilet";
-import FeatureNameHeader from "~/needs-refactoring/components/CombinedFeaturePanel/components/FeatureNameHeader";
-import type { BaseEditorProps } from "./BaseEditor";
 import { PrimaryButton } from "~/components/button/PrimaryButton";
 import { SecondaryButton } from "~/components/button/SecondaryButton";
 import { StyledReportView } from "~/needs-refactoring/components/CombinedFeaturePanel/ReportView";
-import ToiletRadioCards from "~/modules/edit/components/ToiletRadioCards";
+import FeatureNameHeader from "~/needs-refactoring/components/CombinedFeaturePanel/components/FeatureNameHeader";
+import type { YesNoUnknown } from "~/needs-refactoring/lib/model/ac/Feature";
+import { isOrHasAccessibleToilet } from "~/needs-refactoring/lib/model/accessibility/isOrHasAccessibleToilet";
+import ToiletRadioCards from "~/pages/[placeType]/[id]/_components/ToiletRadioCards";
+import type { BaseEditorProps } from "./BaseEditor";
 
 export const ToiletsWheelchairEditor: React.FC<BaseEditorProps> = ({
   feature,
