+++ src/pages/[placeType]/[id]/_hooks/useNextAccessibleToilet.ts
