+++ src/pages/[placeType]/[id]/_tests/add-wheelchair-description.e2e-spec.ts
@@ -1,5 +1,5 @@
 import { expect, type Locator, test } from "@playwright/test";
-import node4544823443Mock from "~/modules/edit/tests/mocks/node-4544823443-osm-mock.json";
+import node4544823443Mock from "~/pages/[placeType]/[id]/_tests/mocks/node-4544823443-osm-mock.json";
 import {
   getButton,
   getDialog,
@@ -7,7 +7,7 @@ import {
   getMenuItem,
   selectLanguage,
   setupPage,
-} from "~/modules/edit/tests/utils";
+} from "~/pages/[placeType]/[id]/_tests/utils";
 
 test.describe("Add wheelchair description in new language", () => {
   let dialog: Locator;
