+++ src/pages/[placeType]/[id]/_tests/mocks/place-infos-mock.json
