+++ src/pages/[placeType]/[id]/_tests/mocks/way-126125230-osm-mock.json
