+++ src/pages/[placeType]/[id]/_utils/attachPropsRecursively.ts
@@ -1,9 +1,9 @@
-import { getTagValues } from "~/modules/feature-panel/utils/getTagValues";
 import { valueRenderFunctions } from "~/needs-refactoring/components/CombinedFeaturePanel/components/AccessibilitySection/valueRenderFunctions";
-import { getOSMTagProps } from "~/needs-refactoring/lib/model/osm/tag-config/getOSMTagProps";
-import type { TagOrTagGroup } from "~/modules/feature-panel/hooks/useOsmTags";
-import type { AnyFeature } from "~/needs-refactoring/lib/model/geo/AnyFeature";
 import type IAccessibilityAttribute from "~/needs-refactoring/lib/model/ac/IAccessibilityAttribute";
+import type { AnyFeature } from "~/needs-refactoring/lib/model/geo/AnyFeature";
+import { getOSMTagProps } from "~/needs-refactoring/lib/model/osm/tag-config/getOSMTagProps";
+import type { TagOrTagGroup } from "~/pages/[placeType]/[id]/_hooks/useOsmTags";
+import { getTagValues } from "~/pages/[placeType]/[id]/_utils/getTagValues";
 
 export function attachTagPropsRecursively(
   tagOrGroup: TagOrTagGroup,
