+++ src/pages/[placeType]/[id]/_utils/generateTree.ts
