--- tests/e2e/specs-todo/different-place-types.js
@@ -1,70 +1,0 @@
-describe("Different place-of-interedt types:", () => {
-  describe("an OpenStreetMap place", () => {
-    it("has a title");
-    it("has an icon");
-    it("shows a green icon");
-    it("shows a yellow icon");
-    it("shows a red icon");
-    it("has a wheelchair accessibility editor");
-    it("has a source link going to OpenStreetMap");
-    it("has an image gallery");
-    it("has an address with a map link");
-    it("has a share link");
-    it("has a report link");
-  });
-
-  describe("a place from Jaccede (loaded over accessibility.cloud)", () => {
-    // /nodes/HNWTJNGmRPBazeQom
-  });
-
-  describe("a place from AXSMap", () => {
-    // /nodes/cxfMJhsZ88bWZeXtW?lat=40.7306&lon=-73.9866&zoom=16
-    // /nodes/Y5dr3q8rGDDoSC82f
-    // /nodes/75KTF4SKDuYWgHRoE
-  });
-
-  describe("a place from HERE", () => {});
-
-  describe("a place from Wheelmap’s accessibility.cloud data source", () => {});
-
-  describe("an elevator", () => {});
-
-  describe("an escalator", () => {});
-
-  describe("an escalator", () => {});
-
-  describe("a place with nice photos", () => {});
-
-  describe("a place without a name", () => {
-    it("displays its category as name");
-    // http://localhost:3000/beta/nodes/-486560173
-  });
-});
-
-// Place from 'Reisen für alle':
-// http://localhost:3000/beta/nodes/wtQSxF2n4fZAKpTdJ
-
-// Places with nice photos:
-// http://localhost:3000/beta/nodes/619366460?zoom=18&lat=52.5147&lon=13.3905
-// http://localhost:3000/beta/nodes/318526010
-
-// Places with ratings from AXSMap:
-
-// Place from FourSquare:
-// http://localhost:3000/beta/nodes/HM7RQjBYvCFb3X4Lk
-
-// Place from HERE Maps:
-// http://localhost:3000/beta/nodes/i82dZGDr7CfvYxE6T
-
-// Accessible parking lot from Parkopedia:
-// http://localhost:3000/beta/nodes/9HSxP37boGZgDBNrd
-
-// An elevator in Mettmann, Germany:
-// http://localhost:3000/beta/nodes/AjgDZDAJKf3wHZXNl/equipment/d9KwaFwaRJk2pjxCr?lat=51.2487&lon=6.9798&zoom=18
-
-// Elevator with probably unknown status:
-// http://localhost:3000/beta/nodes/AjgDZDAJKf3wHZXNl/equipment/MT3iPeh7QHeoXptL5?lat=51.2487&lon=6.9798&zoom=18
-
-// A station, often has elevator/escalator disruptions:
-// http://localhost:3000/nodes/hBkEGMEzkJ659qQwG
-// });
